Change a reconnect test to be more meaningful
The previously tested behavior of client reconnecting to host is less guaranteed than the test assumed; since client is not sending any further rpcs to host, it may not detect that the connection is dead (implementation and backend dependent) and the test would fail.
It makes more sense for client to call into host, in which case a reconnect is guaranteed (since client was the one connecting to host in the first place).