Memfd allocator.
This introduces a memory allocator based on the memfd_create linux system call.
memfd_create creates an in-memory file, which can be mmapped and shared between processes. This file is automatically deleted (and memory freed) when the last process using it exits.
This will be the basis of future faster shared memory transports, and gives us firmer control over memory (eg for memory locking, communicating with cuda/infiniband)