Memfd allocator (#41)
* Removed tensorpipe submodule

* Remove references to tensorpipe

* clang-format

* BufferHandle arguments/returns, speed-up calls.

This allows functions to take and return BufferHandle. These must be used with deserializeArguments and serializeReturn.

This allows python functions to deserialize, call, and serialize return values all under the same GIL, and allows us to drop the use of GilWrapper in those places.

* move buffer.h and allocator.h into memory folder

* Memfd allocator.

This introduces a memory allocator based on the memfd_create linux system call.
memfd_create creates an in-memory file, which can be mmapped and shared between processes. This file is automatically deleted (and memory freed) when the last process using it exits.
This will be the basis of future faster shared memory transports, and gives us firmer control over memory (eg for memory locking, communicating with cuda/infiniband)