Improved peer discovery & connection logic.
We now resolve our local interface addresses (when listening on any address), to use in greetings to other peers.
We also preserve all original addresses in onGreeting, in addition to the existing remote address + port fiddling.
These together fixes an issue where if A and B are connected through a loopback interface, and C connects to A, then asks A for information about B, C would only receive loopback addresses, which may not be usable.

In addition, there is now some additional logic around connecting which uses boot_id (read from /proc) to determine if we should even try to connect through loopback addresses or through unix sockets.
This makes establishing a working connection quicker.